converted "Change Password" checkbox to Angular toggle button/link.
This removes the support for non-JavaScript use of the page, and completes the re-implementation of the previous  Prototype behavior of this page in Angular.  However, this still keeps the same basic flow, with most validation and errors done on the server side, so not very Angularish yet (e.g., still using a regular form submit, not AJAX).