added applicationContext.xml to git (stopped ignoring).
Heroku complained that it was missing.  I had a similar problem when I ran locally last month, I think.  I am not sure if Grails is supposed to create that file when it runs, or IntelliJ copied it from somewhere, or what, but I'm not really worrying about that now, just trying to get it running.