When not editing, changed disabled form controls to static text.
This avoids showing the user disabled form input controls, which are confusing, especially when seen for the first time without comparison to enabled input controls.  The "Edit" button is enough of an indication that  something on the page can be edited; the user can easily try it to see what is editable.

This change also adds ng-cloak for slower browsers, although Chrome is too fast for me to see any problem.