update Angular-UI bootstrap from 0.10.0 to 0.11.0.
The update includes this breaking change:  https://github.com/angular-ui/bootstrap/pull/1460  so I had to change the alert type attribute to interpolated.  I made the same change to bendy-alert-icon for consistency.

I did not include the update of the template-less versions, because I'm not sure if I am going to use them.