added icons for phone and connection type.
Also stopped displaying the level.  I want to distinguish a direct line at an office from a general phone number, but I don't know how I can do that in icons without confusing the users.  I think it might be better to have the users who are interested in that distinction to expand the details of the contact to see that.