added basic sorting.
There is still an issue with nulls:  rows with selected sort properties that are all null seem to be excluded from the results.  I don't know if this is an issue with Grails, Hibernate, or H2 (which I'm just using for development).  Hoping to work around it with Hibernate 4, based on https://hibernate.atlassian.net/browse/HHH-465, I tried updating from Hibernate 3 and configuring hibernate.order_by.default_null_ordering, but that didn't help (and Hibernate 4 generated additional warnings on startup about the schema and foreign keys, so I stayed on Hibernate 3).  I won't worry about this anymore for now.