added Heroku webapp-runner 7.0.40.0.
That's the version that Heroku used by default.  Heroku's web site says it runs the app in Tomcat.  It's the latest version on http://repo2.maven.org/maven2/com/github/jsimone/webapp-runner/ but unfortunately a year old.  Assuming it corresponds to the Tomcat version, I wanted to try the Tomcat that the app is currently configured to use, 7.0.52.1, for proper compatibility with this version of Grails (2.3.7).

By adding this old one to the app, I can at least use it to run locally, to try to reproduce the problem I'm seeing now on Heroku (i.e., that the people.json URL does not map to the controller's index method and redirects to the non-RESTful URL).  Unfortunately, I did not see that problem locally, so I could not debug it.  But, I'll leave the webapp-runner in the project, in case I want to try running it locally again, or try to build a current version from https://github.com/jsimone/webapp-runner.