changed prod db to "create-drop".
I needed to do this to run the war in webapp-runner more than once locally.  Since the prod db is configured as a file, not mem, but the fixtures still try to load on bootstrap (because I need them to test on Heroku), the boot fails with duplicate errors.  So, for now, I'll just drop the prod db.