added initial search using ElasticSearch.
This includes full text search w/ sorting and paging, but not highlighting, and still need to add/merge unindexed fields into results.  Also upgraded Grails from 2.3.6 to 2.3.11.

I initially tried the searchable plugin, since it seemed the more popular, canonical, and better documented.  However, I had problems with it, and it also seemed to be obsolete and abandoned, based on Compass.  ElasticSearch seems a lot more alive, although they're both based on Lucene, and the plugin docs are sparse.